Added payment

+++ src/com/fitness/mantra/model/Plan.java
@@ -13,12 +13,11 @@ public class Plan {
 
 	private Integer id;
 	private String name;
-	private Integer price;
+	private Double price;
 
 	public Plan(HttpServletRequest request) {
 		this.id = request.getParameter("id") != null ? Integer.parseInt(request.getParameter("id")) : null;
 		this.name = request.getParameter("name");
-		this.price = request.getParameter("price") != null ? Integer.parseInt(request.getParameter("price")) : null;
+		this.price = request.getParameter("price") != null ? Double.parseDouble(request.getParameter("price")) : null;
 	}
-
 }
